+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 Features:
 
  - "Extension updated" links to the changelog
- - Download current url when no link or text selected
+ - Download current page's url when no link or text selected
  - Notification when reached 75% of quota for limited hosters
  - Notification when only 7 days left of premium
 
@@ -11,3 +11,4 @@ Bugfixes:
 
  - Persistent storage of warnings
  - Handle case where account information is requested when not logged in
+ - No "extension updated" notification in case of bugfix